refactor(tests): split monolithic EdutokTests.swift into 11 domain files
All 68 @Test functions preserved, now organized by domain:
LevelingTests, TopicProgressTests, StreakTests, LeaderboardTests,
JSONParsingTests, SpacedRepetitionTests, XPMathTests, ChallengeTests,
AchievementTests, GeminiClientTests, CalendarTests.
